msa_build_files/requirements.txt
PySide6>=6.5.0
peewee>=3.17.0
qrcode>=7.4.2
python-dotenv>=1.0.0
psutil>=5.9.0; sys_platform == 'win32'
uuid>=1.30
absl-py>=1.4.0
python-Levenshtein>=0.21.0
python-dateutil>=2.8.2
Pillow>=10.0.0
matplotlib>=3.5.0
phonenumbers>=8.13.0
cryptography>=41.0.0
pywin32; sys_platform == 'win32'
fpdf
weasyprint>=60.1
python-barcode>=0.14.0
